pyproject.toml
[tool.poetry]
name = "imkg-wiki-llm"
version = "0.1.0"
description = ""
authors = ["Harry <141598639+harrytran001@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tiktoken = "^0.8.0"
rdflib = "^7.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
